# openapi.yaml
openapi: 3.0.3
info:
  title: HabitOS API
  version: 1.0.0
  description: API for HabitOS - Behavioral Optimization Platform
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - timestamp

    # Auth Schemas
...
    BehaviorCategory:
      type: string
      enum: [health, productivity, learning, social, financial, creativity, mindfulness]
    TimeSlot:
      type: string
      enum: [early_morning, morning, midday, afternoon, evening, night]
    Objective:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 10
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - description
        - priority
        - createdAt
    ObjectiveImpact:
      type: object
      properties:
        objectiveId:
          type: string
        objectiveName:
          type: string
        impactScore:
          type: number
          description: Score from -1 to 1
      required:
        - objectiveId
        - objectiveName
        - impactScore
    Behavior:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/BehaviorCategory'
        energyCost:
          type: integer
          minimum: 1
          maximum: 10
        durationMin:
          type: integer
        durationMax:
          type: integer
        preferredTimeSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
        objectiveImpacts:
          type: array
          items:
            $ref: '#/components/schemas/ObjectiveImpact'
        isActive:
          type: boolean
        frequency:
          type: string
          enum: [daily, weekly, custom]
        frequencyCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - description
        - category
        - energyCost
        - durationMin
        - durationMax
        - preferredTimeSlots
        - objectiveImpacts
        - isActive
        - frequency
        - createdAt
        - updatedAt
    BehaviorFormData:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/BehaviorCategory'
        energyCost:
          type: integer
        durationMin:
          type: integer
        durationMax:
          type: integer
        preferredTimeSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
        objectiveImpacts:
          type: array
          items:
            $ref: '#/components/schemas/ObjectiveImpact'
        isActive:
          type: boolean
        frequency:
          type: string
          enum: [daily, weekly, custom]
        frequencyCount:
          type: integer
      required:
        - name
        - description
        - category
        - energyCost
        - durationMin
        - durationMax
        - preferredTimeSlots
        - objectiveImpacts
        - isActive
        - frequency

    # Optimization Schemas
...
paths:
...
  /behaviors:
    get:
      summary: Get user behaviors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of behaviors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Behavior'
    post:
      summary: Create behavior
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BehaviorFormData'
      responses:
        '201':
          description: Behavior created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Behavior'
  /behaviors/{id}:
    get:
      summary: Get behavior details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Behavior details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Behavior'
    put:
      summary: Update behavior
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BehaviorFormData'
      responses:
        '200':
          description: Behavior updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Behavior'
    delete:
      summary: Delete behavior
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Behavior deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
  /behaviors/objectives:
    get:
      summary: Get user objectives
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of objectives
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Objective'

  # Optimization
...